Fix 2 naming issues.
