+++ apm-sniffer/apm-toolkit-activation/apm-toolkit-trace-activation/src/main/java/org/apache/skywalking/apm/toolkit/activation/ToolkitPluginConfig.java
@@ -16,25 +16,18 @@
  *
  */
 
-package org.apache.skywalking.apm.plugin.customize.conf;
+package org.apache.skywalking.apm.toolkit.activation;
 
-import java.util.HashMap;
-import java.util.Map;
 import org.apache.skywalking.apm.agent.core.boot.PluginConfig;
 
-public class PluginConfigService implements PluginConfig {
+public class ToolkitPluginConfig implements PluginConfig {
     public static class Plugin {
-        public static class Customize {
+        public static class Toolkit {
             /**
-             * Custom enhancement class configuration file path, recommended to use an absolute path.
+             * If true, the fully qualified method name will be used as the operation name instead of the given
+             * operation name, default is false.
              */
-            public static String ENHANCE_FILE = "";
-
-            /**
-             * Some information after custom enhancements, this configuration is used by the custom enhancement plugin.
-             * And using Map CONTEXT for avoiding classloader isolation issue.
-             */
-            public static Map<String, Object> CONTEXT = new HashMap<>();
+            public static boolean USE_QUALIFIED_NAME_AS_OPERATION_NAME = false;
         }
     }
 }
